fix(org-people): address PR #945 follow-up review threads (LFXV2-1677)
- Mark any live-only person (no `snowflake` source) as synthetic, not just
  access-only rows, so board/committee/key-contact rows added before the next
  dbt build don't render an expandable chevron that fires a Snowflake detail
  fetch with a `live-` key that matches nothing.
- Bound the OrgPeopleDirectoryService singleton cache at 2,000 entries with
  oldest-first eviction, mirroring OrgMembershipResolverService, so it can't
  grow unbounded across every (org x caller) pair over the server's lifetime.
- Extract the duplicated `splitName` helper into a shared `splitDisplayName`
  util in @lfx-one/shared and consume it from both org services.
- Replace the imperative broken-avatar DOM mutation with a declarative
  failed-avatar signal, avoiding direct DOM access in a zoneless component.
